Removed V1 pydantic model generation and references completely from the project, renaming the pydantic V2 model to the simple "pydanticmodel.py" file name (fixed the makefile accordingly); poetry dependency for pydantic itself enforces "^2"
