+++ Makefile
@@ -108,7 +108,7 @@ gen-examples:
 
 gen-project: $(PYMODEL)
 	$(RUN) gen-project ${GEN_PARGS} -d $(DEST) $(SOURCE_SCHEMA_PATH) && mv $(DEST)/*.py $(PYMODEL)
-	$(RUN) gen-pydantic --pydantic_version 2 src/translator_testing_model/schema/translator_testing_model.yaml > $(PYMODEL)/pydanticmodel_v2.py
+	$(RUN) gen-pydantic --pydantic-version 2 src/translator_testing_model/schema/translator_testing_model.yaml > $(PYMODEL)/pydanticmodel.py
 
 
 test: test-schema test-python test-examples
