+++ README.md
@@ -18,8 +18,7 @@ Capturing these details in metadata that is parsable and usable by test runners 
       (edit this)
     * [datamodel](src/translator_testing_model/datamodel/README.md) -- generated
       Python datamodels
-      * [Pydantic V1](src/translator_testing_model/datamodel/pydanticmodel.py) - the version 1 model of Pydantic is now deprecated! Use V2.
-      * [Pydantic V2](src/translator_testing_model/datamodel/pydanticmodel_v2.py)
+      * [Pydantic](src/translator_testing_model/datamodel/pydanticmodel.py) - this is a version 2 model.
       * [Python Dataclasses](src/translator_testing_model/datamodel/translator_testing_model.py)
 * [tests/](tests/test_data.py) - Python tests
