+++ src/translator_testing_model/datamodel/README.md
@@ -1,5 +1,4 @@
 # Generated Python Datamodels
 
-* [Pydantic V1](src/translator_testing_model/datamodel/pydanticmodel.py) - the version 1 model of Pydantic is now deprecated! Use V2.
-* [Pydantic V2](src/translator_testing_model/datamodel/pydanticmodel_v2.py)
+* [Pydantic](src/translator_testing_model/datamodel/pydanticmodel.py) -  this is a version 2 model
 * [Python Dataclasses](src/translator_testing_model/datamodel/translator_testing_model.py)
