+++ src/translator_testing_model/datamodel/pydanticmodel.py
@@ -1,33 +1,52 @@
-from __future__ import annotations
-from datetime import datetime, date
-from enum import Enum
-from typing import List, Dict, Optional, Any, Union
-from pydantic import BaseModel as BaseModel, ConfigDict, Field
+from __future__ import annotations 
+from datetime import (
+    datetime,
+    date
+)
+from decimal import Decimal 
+from enum import Enum 
+import re
 import sys
-if sys.version_info >= (3, 8):
-    from typing import Literal
+from typing import (
+    Any,
+    List,
+    Literal,
+    Dict,
+    Optional,
+    Union
+)
+from pydantic.version import VERSION  as PYDANTIC_VERSION 
+if int(PYDANTIC_VERSION[0])>=2:
+    from pydantic import (
+        BaseModel,
+        ConfigDict,
+        Field,
+        field_validator
+    )
 else:
-    from typing_extensions import Literal
-
+    from pydantic import (
+        BaseModel,
+        Field,
+        validator
+    )
 
 metamodel_version = "None"
 version = "0.0.0"
 
-class WeakRefShimBaseModel(BaseModel):
-   __slots__ = '__weakref__'
 
-class ConfiguredBaseModel(WeakRefShimBaseModel,
-                validate_assignment = True,
-                validate_all = True,
-                underscore_attrs_are_private = True,
-                extra = 'forbid',
-                arbitrary_types_allowed = True,
-                use_enum_values = True):
+class ConfiguredBaseModel(BaseModel):
+    model_config = ConfigDict(
+        validate_assignment = True,
+        validate_default = True,
+        extra = "forbid",
+        arbitrary_types_allowed = True,
+        use_enum_values = True,
+        strict = False,
+    )
     pass
 
 
 class TestSourceEnum(str, Enum):
-    
     # (External) Subject Matter Expert
     SME = "SME"
     # Subject Matter User Reasonably Familiar, generally Translator-internal biomedical science expert
@@ -42,23 +61,20 @@ class TestSourceEnum(str, Enum):
     TranslatorTeam = "TranslatorTeam"
     # Current SRI_Testing-like test data edges specific to KP or ARA components
     TestDataLocation = "TestDataLocation"
-    
-    
+
 
 class TestObjectiveEnum(str, Enum):
-    
     # Acceptance (pass/fail) test
     AcceptanceTest = "AcceptanceTest"
     # Semantic benchmarking
     BenchmarkTest = "BenchmarkTest"
     # Quantitative test
     QuantitativeTest = "QuantitativeTest"
-    # One Hop Test of knowledge graph navigation
-    OneHopTest = "OneHopTest"
-    # TRAPI and Biolink Model ("reasoner-validator") validation
+    # Release-specific TRAPI and Biolink Model ("reasoner-validator") compliance validation
     StandardsValidationTest = "StandardsValidationTest"
-    
-    
+    # Knowledge graph "One Hop" query navigation integrity
+    OneHopTest = "OneHopTest"
+
 
 class TestEnvEnum(str, Enum):
     """
@@ -72,77 +88,51 @@ class TestEnvEnum(str, Enum):
     test = "test"
     # Production
     prod = "prod"
-    
-    
+
 
 class FileFormatEnum(str, Enum):
     """
     Text file formats for test data sources.
     """
-    
     TSV = "TSV"
-    
     YAML = "YAML"
-    
     JSON = "JSON"
-    
-    
+
 
 class ExpectedOutputEnum(str, Enum):
     """
     Expected output values for instances of Test Asset or Test Cases(?). (Note: does this Enum overlap with 'ExpectedResultsEnum' below?)
     """
-    
     Acceptable = "Acceptable"
-    
     BadButForgivable = "BadButForgivable"
-    
     NeverShow = "NeverShow"
-    
     TopAnswer = "TopAnswer"
-    
     OverlyGeneric = "OverlyGeneric"
-    
-    
+
 
 class TestIssueEnum(str, Enum):
-    
-    
     causes_not_treats = "causes not treats"
     # 'Text Mining Knowledge Provider' generated relationship?
     TMKP = "TMKP"
-    
     category_too_generic = "category too generic"
-    
     contraindications = "contraindications"
-    
     chemical_roles = "chemical roles"
-    
     test_issue = "test_issue"
-    
-    
+
 
 class SemanticSeverityEnum(str, Enum):
     """
     From Jenn's worksheet, empty or ill defined (needs elaboration)
     """
-    
     High = "High"
-    
     Low = "Low"
-    
     NotApplicable = "NotApplicable"
-    
-    
+
 
 class DirectionEnum(str, Enum):
-    
-    
     increased = "increased"
-    
     decreased = "decreased"
-    
-    
+
 
 class ExpectedResultsEnum(str, Enum):
     """
@@ -152,57 +142,42 @@ class ExpectedResultsEnum(str, Enum):
     include_good = "include_good"
     # The query should not return the result in this test case
     exclude_bad = "exclude_bad"
-    
-    
+
 
 class NodeEnum(str, Enum):
     """
     Target node of a Subject-Predicate-Object driven query
     """
-    
     subject = "subject"
-    
     object = "object"
-    
-    
+
 
 class QueryTypeEnum(str, Enum):
     """
     Query
     """
-    
     treats = "treats"
-    
-    
+
 
 class TrapiTemplateEnum(str, Enum):
-    
-    
     ameliorates = "ameliorates"
-    
     treats = "treats"
-    
     three_hop = "three_hop"
-    
     drug_treats_rare_disease = "drug_treats_rare_disease"
-    
     drug_to_gene = "drug-to-gene"
-    
-    
+
 
 class ComponentEnum(str):
     """
     Translator components are identified by their InfoRes identifiers.
     """
-    
-    dummy = "dummy"
-    
+    pass
+
 
 class TestPersonaEnum(str, Enum):
     """
     User persona context of a given test.
     """
-    
     All = "All"
     # An MD or someone working in the clinical field.
     Clinical = "Clinical"
@@ -219,8 +194,7 @@ class TestCaseResultEnum(str, Enum):
     FAILED = "FAILED"
     # test case result indicating that the specified test was not run.
     SKIPPED = "SKIPPED"
-    
-    
+
 
 class TestEntityParameter(ConfiguredBaseModel):
     """
@@ -228,13 +202,12 @@ class TestEntityParameter(ConfiguredBaseModel):
     """
     parameter: Optional[str] = Field(None, description="""Name of a TestParameter.""")
     value: Optional[str] = Field(None, description="""(String) value of a TestParameter.""")
-    
+
 
 class Qualifier(TestEntityParameter):
-    
     parameter: Optional[str] = Field(None, description="""The 'parameter' of a Qualifier should be a `qualifier` slot name from the Biolink Model ('biolink' namespace) 'biolink:qualifier' hierarchy.""")
     value: Optional[str] = Field(None, description="""The 'value' of should be a suitable value generally drawn from an applicable Biolink Model (\"Enum\") value set of the specified Qualifier.""")
-    
+
 
 class TestEntity(ConfiguredBaseModel):
     """
@@ -245,7 +218,7 @@ class TestEntity(ConfiguredBaseModel):
     description: Optional[str] = Field(None, description="""A human-readable description for a Test Entity""")
     tags: Optional[List[str]] = Field(default_factory=list, description="""A human-readable tags for categorical memberships of a TestEntity (preferably a URI or CURIE). Typically used to aggregate instances of TestEntity into formally typed or ad hoc lists.""")
     test_runner_settings: Optional[List[str]] = Field(default_factory=list, description="""Scalar parameters for the TestRunner processing a given TestEntity.""")
-    
+
 
 class TestMetadata(TestEntity):
     """
@@ -253,14 +226,14 @@ class TestMetadata(TestEntity):
     """
     test_source: Optional[TestSourceEnum] = Field(None, description="""Provenance of a specific set of test assets, cases and/or suites.  Or, the person who cares about this,  know about this.  We would like this to be an ORCID eventually, but currently it is just a string.""")
     test_reference: Optional[str] = Field(None, description="""Document URL where original test source particulars are registered (e.g. Github repo)""")
-    test_objective: Optional[TestObjectiveEnum] = Field(None, description="""Testing objective behind specified set of test particulars (e.g. acceptance pass/fail; benchmark; quantitative)""")
+    test_objective: Optional[TestObjectiveEnum] = Field(None, description="""Testing objective behind specified set of test particulars (e.g. acceptance pass/fail; benchmark; quantitative; standards compliance; graph navigation integrity)""")
     test_annotations: Optional[List[TestEntityParameter]] = Field(default_factory=list, description="""Metadata annotation.""")
     id: str = Field(..., description="""A unique identifier for a Test Entity""")
     name: Optional[str] = Field(None, description="""A human-readable name for a Test Entity""")
     description: Optional[str] = Field(None, description="""A human-readable description for a Test Entity""")
     tags: Optional[List[str]] = Field(default_factory=list, description="""A human-readable tags for categorical memberships of a TestEntity (preferably a URI or CURIE). Typically used to aggregate instances of TestEntity into formally typed or ad hoc lists.""")
     test_runner_settings: Optional[List[str]] = Field(default_factory=list, description="""Scalar parameters for the TestRunner processing a given TestEntity.""")
-    
+
 
 class TestAsset(TestEntity):
     """
@@ -288,7 +261,7 @@ class TestAsset(TestEntity):
     description: Optional[str] = Field(None, description="""A human-readable description for a Test Entity""")
     tags: Optional[List[str]] = Field(default_factory=list, description="""One or more 'tags' slot values (inherited from TestEntity) should generally be defined to specify TestAsset membership in a \"Block List\" collection""")
     test_runner_settings: Optional[List[str]] = Field(default_factory=list, description="""Scalar settings for the TestRunner, e.g. \"inferred\"""")
-    
+
 
 class AcceptanceTestAsset(TestAsset):
     """
@@ -326,7 +299,7 @@ class AcceptanceTestAsset(TestAsset):
     description: Optional[str] = Field(None, description="""A human-readable description for a Test Entity""")
     tags: Optional[List[str]] = Field(default_factory=list, description="""One or more 'tags' slot values (inherited from TestEntity) should generally be defined to specify TestAsset membership in a \"Block List\" collection""")
     test_runner_settings: Optional[List[str]] = Field(default_factory=list, description="""Scalar settings for the TestRunner, e.g. \"inferred\"""")
-    
+
 
 class TestEdgeData(TestAsset):
     """
@@ -354,7 +327,7 @@ class TestEdgeData(TestAsset):
     description: Optional[str] = Field(None, description="""A human-readable description for a Test Entity""")
     tags: Optional[List[str]] = Field(default_factory=list, description="""One or more 'tags' slot values (inherited from TestEntity) should generally be defined to specify TestAsset membership in a \"Block List\" collection""")
     test_runner_settings: Optional[List[str]] = Field(default_factory=list, description="""Scalar settings for the TestRunner, e.g. \"inferred\"""")
-    
+
 
 class Precondition(TestEntity):
     """
@@ -365,7 +338,7 @@ class Precondition(TestEntity):
     description: Optional[str] = Field(None, description="""A human-readable description for a Test Entity""")
     tags: Optional[List[str]] = Field(default_factory=list, description="""A human-readable tags for categorical memberships of a TestEntity (preferably a URI or CURIE). Typically used to aggregate instances of TestEntity into formally typed or ad hoc lists.""")
     test_runner_settings: Optional[List[str]] = Field(default_factory=list, description="""Scalar parameters for the TestRunner processing a given TestEntity.""")
-    
+
 
 class TestCase(TestEntity):
     """
@@ -375,7 +348,7 @@ class TestCase(TestEntity):
     test_assets: List[TestAsset] = Field(default_factory=list, description="""One or more 'tags' slot values (inherited from TestEntity) should generally be defined as filters to specify TestAsset membership in 'test_assets' slot (\"Block List\") collection.""")
     preconditions: Optional[List[str]] = Field(default_factory=list)
     trapi_template: Optional[TrapiTemplateEnum] = Field(None, description="""A template for a query, which can be used to generate a query for a test case.  note: the current enumerated values for this slot come from the Benchmarks repo config/benchmarks.json \"templates\" collection and refer to the \"name\" field of each template.  Templates themselves are currently stored in the config/[source_name]/templates directory.""")
-    test_case_objective: Optional[TestObjectiveEnum] = Field(None, description="""Testing objective behind specified set of test particulars (e.g. acceptance pass/fail; benchmark; quantitative)""")
+    test_case_objective: Optional[TestObjectiveEnum] = Field(None, description="""Testing objective behind specified set of test particulars (e.g. acceptance pass/fail; benchmark; quantitative; standards compliance; graph navigation integrity)""")
     test_case_source: Optional[TestSourceEnum] = Field(None, description="""Provenance of a specific set of test assets, cases and/or suites.  Or, the person who cares about this,  know about this.  We would like this to be an ORCID eventually, but currently it is just a string.""")
     test_case_predicate_name: Optional[str] = Field(None)
     test_case_predicate_id: Optional[str] = Field(None)
@@ -390,7 +363,7 @@ class TestCase(TestEntity):
     description: Optional[str] = Field(None, description="""A human-readable description for a Test Entity""")
     tags: Optional[List[str]] = Field(default_factory=list, description="""One or more 'tags' slot values (slot inherited from TestEntity) should generally be defined as filters to specify TestAsset membership in a \"Block List\" collection.""")
     test_runner_settings: Optional[List[str]] = Field(default_factory=list, description="""Scalar parameters for the TestRunner processing a given TestEntity.""")
-    
+
 
 class AcceptanceTestCase(TestCase):
     """
@@ -400,7 +373,7 @@ class AcceptanceTestCase(TestCase):
     test_assets: List[AcceptanceTestAsset] = Field(default_factory=list, description="""One or more 'tags' slot values (inherited from TestEntity) should generally be defined as filters to specify TestAsset membership in 'test_assets' slot (\"Block List\") collection.""")
     preconditions: Optional[List[str]] = Field(default_factory=list)
     trapi_template: Optional[TrapiTemplateEnum] = Field(None, description="""A template for a query, which can be used to generate a query for a test case.  note: the current enumerated values for this slot come from the Benchmarks repo config/benchmarks.json \"templates\" collection and refer to the \"name\" field of each template.  Templates themselves are currently stored in the config/[source_name]/templates directory.""")
-    test_case_objective: Optional[TestObjectiveEnum] = Field(None, description="""Testing objective behind specified set of test particulars (e.g. acceptance pass/fail; benchmark; quantitative)""")
+    test_case_objective: Optional[TestObjectiveEnum] = Field(None, description="""Testing objective behind specified set of test particulars (e.g. acceptance pass/fail; benchmark; quantitative; standards compliance; graph navigation integrity)""")
     test_case_source: Optional[TestSourceEnum] = Field(None, description="""Provenance of a specific set of test assets, cases and/or suites.  Or, the person who cares about this,  know about this.  We would like this to be an ORCID eventually, but currently it is just a string.""")
     test_case_predicate_name: Optional[str] = Field(None)
     test_case_predicate_id: Optional[str] = Field(None)
@@ -415,7 +388,7 @@ class AcceptanceTestCase(TestCase):
     description: Optional[str] = Field(None, description="""A human-readable description for a Test Entity""")
     tags: Optional[List[str]] = Field(default_factory=list, description="""One or more 'tags' slot values (slot inherited from TestEntity) should generally be defined as filters to specify TestAsset membership in a \"Block List\" collection.""")
     test_runner_settings: Optional[List[str]] = Field(default_factory=list, description="""Scalar parameters for the TestRunner processing a given TestEntity.""")
-    
+
 
 class QuantitativeTestCase(TestCase):
     """
@@ -425,7 +398,7 @@ class QuantitativeTestCase(TestCase):
     test_assets: List[TestAsset] = Field(default_factory=list, description="""One or more 'tags' slot values (inherited from TestEntity) should generally be defined as filters to specify TestAsset membership in 'test_assets' slot (\"Block List\") collection.""")
     preconditions: Optional[List[str]] = Field(default_factory=list)
     trapi_template: Optional[TrapiTemplateEnum] = Field(None, description="""A template for a query, which can be used to generate a query for a test case.  note: the current enumerated values for this slot come from the Benchmarks repo config/benchmarks.json \"templates\" collection and refer to the \"name\" field of each template.  Templates themselves are currently stored in the config/[source_name]/templates directory.""")
-    test_case_objective: Optional[TestObjectiveEnum] = Field(None, description="""Testing objective behind specified set of test particulars (e.g. acceptance pass/fail; benchmark; quantitative)""")
+    test_case_objective: Optional[TestObjectiveEnum] = Field(None, description="""Testing objective behind specified set of test particulars (e.g. acceptance pass/fail; benchmark; quantitative; standards compliance; graph navigation integrity)""")
     test_case_source: Optional[TestSourceEnum] = Field(None, description="""Provenance of a specific set of test assets, cases and/or suites.  Or, the person who cares about this,  know about this.  We would like this to be an ORCID eventually, but currently it is just a string.""")
     test_case_predicate_name: Optional[str] = Field(None)
     test_case_predicate_id: Optional[str] = Field(None)
@@ -440,7 +413,7 @@ class QuantitativeTestCase(TestCase):
     description: Optional[str] = Field(None, description="""A human-readable description for a Test Entity""")
     tags: Optional[List[str]] = Field(default_factory=list, description="""One or more 'tags' slot values (slot inherited from TestEntity) should generally be defined as filters to specify TestAsset membership in a \"Block List\" collection.""")
     test_runner_settings: Optional[List[str]] = Field(default_factory=list, description="""Scalar parameters for the TestRunner processing a given TestEntity.""")
-    
+
 
 class TestSuiteSpecification(TestEntity):
     """
@@ -453,7 +426,7 @@ class TestSuiteSpecification(TestEntity):
     description: Optional[str] = Field(None, description="""A human-readable description for a Test Entity""")
     tags: Optional[List[str]] = Field(default_factory=list, description="""A human-readable tags for categorical memberships of a TestEntity (preferably a URI or CURIE). Typically used to aggregate instances of TestEntity into formally typed or ad hoc lists.""")
     test_runner_settings: Optional[List[str]] = Field(default_factory=list, description="""Scalar parameters for the TestRunner processing a given TestEntity.""")
-    
+
 
 class TestSuite(TestEntity):
     """
@@ -468,10 +441,9 @@ class TestSuite(TestEntity):
     description: Optional[str] = Field(None, description="""A human-readable description for a Test Entity""")
     tags: Optional[List[str]] = Field(default_factory=list, description="""A human-readable tags for categorical memberships of a TestEntity (preferably a URI or CURIE). Typically used to aggregate instances of TestEntity into formally typed or ad hoc lists.""")
     test_runner_settings: Optional[List[str]] = Field(default_factory=list, description="""Scalar parameters for the TestRunner processing a given TestEntity.""")
-    
+
 
 class AcceptanceTestSuite(TestSuite):
-    
     test_metadata: TestMetadata = Field(..., description="""Test metadata describes the external provenance, cross-references and objectives for a given test.""")
     test_persona: Optional[TestPersonaEnum] = Field(None, description="""A Test persona describes the user or operational context of a given test.""")
     test_cases: Optional[Dict[str, TestCase]] = Field(default_factory=dict, description="""List of explicitly enumerated Test Cases.""")
@@ -481,14 +453,14 @@ class AcceptanceTestSuite(TestSuite):
     description: Optional[str] = Field(None, description="""A human-readable description for a Test Entity""")
     tags: Optional[List[str]] = Field(default_factory=list, description="""A human-readable tags for categorical memberships of a TestEntity (preferably a URI or CURIE). Typically used to aggregate instances of TestEntity into formally typed or ad hoc lists.""")
     test_runner_settings: Optional[List[str]] = Field(default_factory=list, description="""Scalar parameters for the TestRunner processing a given TestEntity.""")
-    
+
 
 class BenchmarkTestSuite(ConfiguredBaseModel):
     """
     JsonObj(is_a='TestSuite')
     """
-    None
-    
+    pass
+
 
 class StandardsComplianceTestSuite(TestSuite):
     """
@@ -503,7 +475,7 @@ class StandardsComplianceTestSuite(TestSuite):
     description: Optional[str] = Field(None, description="""A human-readable description for a Test Entity""")
     tags: Optional[List[str]] = Field(default_factory=list, description="""A human-readable tags for categorical memberships of a TestEntity (preferably a URI or CURIE). Typically used to aggregate instances of TestEntity into formally typed or ad hoc lists.""")
     test_runner_settings: Optional[List[str]] = Field(default_factory=list, description="""Scalar parameters for the TestRunner processing a given TestEntity.""")
-    
+
 
 class OneHopTestSuite(TestSuite):
     """
@@ -518,7 +490,7 @@ class OneHopTestSuite(TestSuite):
     description: Optional[str] = Field(None, description="""A human-readable description for a Test Entity""")
     tags: Optional[List[str]] = Field(default_factory=list, description="""A human-readable tags for categorical memberships of a TestEntity (preferably a URI or CURIE). Typically used to aggregate instances of TestEntity into formally typed or ad hoc lists.""")
     test_runner_settings: Optional[List[str]] = Field(default_factory=list, description="""Scalar parameters for the TestRunner processing a given TestEntity.""")
-    
+
 
 class TestCaseResult(TestEntity):
     """
@@ -532,7 +504,7 @@ class TestCaseResult(TestEntity):
     description: Optional[str] = Field(None, description="""A human-readable description for a Test Entity""")
     tags: Optional[List[str]] = Field(default_factory=list, description="""A human-readable tags for categorical memberships of a TestEntity (preferably a URI or CURIE). Typically used to aggregate instances of TestEntity into formally typed or ad hoc lists.""")
     test_runner_settings: Optional[List[str]] = Field(default_factory=list, description="""Scalar parameters for the TestRunner processing a given TestEntity.""")
-    
+
 
 class TestRunSession(TestEntity):
     """
@@ -550,7 +522,7 @@ class TestRunSession(TestEntity):
     description: Optional[str] = Field(None, description="""A human-readable description for a Test Entity""")
     tags: Optional[List[str]] = Field(default_factory=list, description="""A human-readable tags for categorical memberships of a TestEntity (preferably a URI or CURIE). Typically used to aggregate instances of TestEntity into formally typed or ad hoc lists.""")
     test_runner_settings: Optional[List[str]] = Field(default_factory=list, description="""Scalar parameters for the TestRunner processing a given TestEntity.""")
-    
+
 
 class TestOutput(TestEntity):
     """
@@ -563,7 +535,7 @@ class TestOutput(TestEntity):
     description: Optional[str] = Field(None, description="""A human-readable description for a Test Entity""")
     tags: Optional[List[str]] = Field(default_factory=list, description="""A human-readable tags for categorical memberships of a TestEntity (preferably a URI or CURIE). Typically used to aggregate instances of TestEntity into formally typed or ad hoc lists.""")
     test_runner_settings: Optional[List[str]] = Field(default_factory=list, description="""Scalar parameters for the TestRunner processing a given TestEntity.""")
-    
+
 
 class TestResultPKSet(TestEntity):
     """
@@ -581,30 +553,29 @@ class TestResultPKSet(TestEntity):
     description: Optional[str] = Field(None, description="""A human-readable description for a Test Entity""")
     tags: Optional[List[str]] = Field(default_factory=list, description="""A human-readable tags for categorical memberships of a TestEntity (preferably a URI or CURIE). Typically used to aggregate instances of TestEntity into formally typed or ad hoc lists.""")
     test_runner_settings: Optional[List[str]] = Field(default_factory=list, description="""Scalar parameters for the TestRunner processing a given TestEntity.""")
-    
 
 
-# Update forward refs
-# see https://pydantic-docs.helpmanual.io/usage/postponed_annotations/
-TestEntityParameter.update_forward_refs()
-Qualifier.update_forward_refs()
-TestEntity.update_forward_refs()
-TestMetadata.update_forward_refs()
-TestAsset.update_forward_refs()
-AcceptanceTestAsset.update_forward_refs()
-TestEdgeData.update_forward_refs()
-Precondition.update_forward_refs()
-TestCase.update_forward_refs()
-AcceptanceTestCase.update_forward_refs()
-QuantitativeTestCase.update_forward_refs()
-TestSuiteSpecification.update_forward_refs()
-TestSuite.update_forward_refs()
-AcceptanceTestSuite.update_forward_refs()
-BenchmarkTestSuite.update_forward_refs()
-StandardsComplianceTestSuite.update_forward_refs()
-OneHopTestSuite.update_forward_refs()
-TestCaseResult.update_forward_refs()
-TestRunSession.update_forward_refs()
-TestOutput.update_forward_refs()
-TestResultPKSet.update_forward_refs()
+# Model rebuild
+# see https://pydantic-docs.helpmanual.io/usage/models/#rebuilding-a-model
+TestEntityParameter.model_rebuild()
+Qualifier.model_rebuild()
+TestEntity.model_rebuild()
+TestMetadata.model_rebuild()
+TestAsset.model_rebuild()
+AcceptanceTestAsset.model_rebuild()
+TestEdgeData.model_rebuild()
+Precondition.model_rebuild()
+TestCase.model_rebuild()
+AcceptanceTestCase.model_rebuild()
+QuantitativeTestCase.model_rebuild()
+TestSuiteSpecification.model_rebuild()
+TestSuite.model_rebuild()
+AcceptanceTestSuite.model_rebuild()
+BenchmarkTestSuite.model_rebuild()
+StandardsComplianceTestSuite.model_rebuild()
+OneHopTestSuite.model_rebuild()
+TestCaseResult.model_rebuild()
+TestRunSession.model_rebuild()
+TestOutput.model_rebuild()
+TestResultPKSet.model_rebuild()
